Moved token files into their own subfolder
