+++ src/app/src/main/kotlin/org/imperial/mrc/hint/AppProperties.kt
@@ -12,10 +12,10 @@ interface AppProperties {
 }
 
 @Configuration
-open class ConfiguredAppProperties: AppProperties {
+open class ConfiguredAppProperties(props: Properties = properties): AppProperties {
 
-    override val uploadDirectory = properties["upload_dir"].toString()
-    override val tokenIssuer = properties["token_issuer"].toString()
+    override val uploadDirectory = props["upload_dir"].toString()
+    override val tokenIssuer = props["token_issuer"].toString()
 
     companion object
     {
@@ -30,7 +30,8 @@ open class ConfiguredAppProperties: AppProperties {
                 }
             }
         }
-        val configPath = "/etc/hint/config.properties"
+
+        var configPath = "/etc/hint/config.properties"
         val properties = readProperties(configPath)
     }
 }
