+++ src/app/src/test/kotlin/org/imperial/mrc/hint/unit/AppPropertiesTest.kt
@@ -1,11 +1,11 @@
 package org.imperial.mrc.hint.unit
 
 import org.assertj.core.api.Java6Assertions.assertThat
-import org.imperial.mrc.hint.AppProperties
 import org.imperial.mrc.hint.ConfiguredAppProperties
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import java.io.File
+import java.util.*
 
 class AppPropertiesTests {
 
@@ -14,18 +14,18 @@ class AppPropertiesTests {
         File("tmp").deleteRecursively()
     }
 
-    private fun makeTempConfigFile(): File {
+    private fun readPropsFromTempFile(contents: String): Properties {
         File("tmp").mkdir()
         val config = File("tmp/fake.properties")
         config.createNewFile()
-        return config
+        config.writeText(contents)
+        return ConfiguredAppProperties.readProperties("tmp/fake.properties")
     }
 
     @Test
     fun `properties can be read from disk`() {
-        val config = makeTempConfigFile()
-        config.writeText("something=1234")
-        val props = ConfiguredAppProperties.readProperties("tmp/fake.properties")
+
+        val props = readPropsFromTempFile("something=1234")
         assertThat(props["something"]).isEqualTo("1234")
     }
 
@@ -35,18 +35,16 @@ class AppPropertiesTests {
     }
 
     @Test
-    fun `can read uploadDirectory`{
-        val config = makeTempConfigFile()
-        config.writeText("upload_dir=/fakedir")
-
-        assertThat((ConfiguredAppProperties as AppProperties).uploadDirectory).isEqualTo("/fakedir")
+    fun `can read uploadDirectory`(){
+        val props = readPropsFromTempFile("upload_dir=/fakedir")
+        val sut = ConfiguredAppProperties(props)
+        assertThat(sut.uploadDirectory).isEqualTo("/fakedir")
     }
 
     @Test
-    fun `can read tokenIssuer`{
-        val config = makeTempConfigFile()
-        config.writeText("token_issuer=fakeIssuer")
-
-        assertThat((ConfiguredAppProperties as AppProperties).tokenIssuer).isEqualTo("fakeIssuer")
+    fun `can read tokenIssuer`(){
+        val props = readPropsFromTempFile("token_issuer=fakeIssuer")
+        val sut = ConfiguredAppProperties(props)
+        assertThat(sut.tokenIssuer).isEqualTo("fakeIssuer")
     }
 }
